requirements.txt
streamlit
langchain
python-dotenv
sqlite3
pandas
tempfile
pydantic
